% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_NAs.R
\name{add_NAs}
\alias{add_NAs}
\title{Add the NA grid cells that may have been removed during the the emulator training process}
\usage{
add_NAs(data, mapping)
}
\arguments{
\item{data}{a matrix of the time by grid cell data that is missing NA gridcells}

\item{mapping}{a mapping data frame created by \code{drop_NAs} that contains the original and new index information for each grid cell.}
}
\value{
a complete data frame of NA grid cells and the generated fields ordered by grid cell
}
\description{
This function adds NA gridcells that were removed during the training process in order to convert
the results to same dimensions as the input grid.
}
\keyword{internal}
