% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{c.griddata}
\alias{c.griddata}
\title{c operator for griddata objects}
\usage{
\method{c}{griddata}(...)
}
\arguments{
\item{...}{One or more griddata objects}
}
\description{
c operator for griddata objects
}
