% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeresiduals.R
\name{characterize.emp.dist}
\alias{characterize.emp.dist}
\title{characterize empirical distributions}
\usage{
characterize.emp.dist(inputresids, len = NULL)
}
\arguments{
\item{inputresids}{Matrix of the original, input residuals for each grid
cell (columns)}

\item{len}{Maximum length of the time series to read.  If the data read is
longer, it will be trimmed.  (Default: read entire time series, regardless of
length.)}
}
\description{
Given a vector of residuals in a single grid cell that empirically
characterizes the distribution of residuals, return a CDF and
quantile function for that grid cell.
}
\details{
The output will be a list with two fields:
\describe{
  \item{\strong{cdf}}{List of empirical CDF functions for each grid cell}
  \item{\strong{quant}}{List of empirical quantile functions for each grid
  cell.}
}



Conventionally, we refer to the output list as \code{empiricaldistribution}.
Notably, any other function with a \code{empiricaldistribution} argument is
expecting one of these structures.
}
