% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTPfullgrids.R
\name{chkinverse}
\alias{chkinverse}
\title{Check if two functions are inverses of one another over a given range of inputs.}
\usage{
chkinverse(f1, f2, range = c(1, 5), ntest = 10, logspace = FALSE)
}
\arguments{
\item{f1}{First function to test}

\item{f2}{Second function to test}

\item{range}{Range of test values}

\item{ntest}{Number of values to test in the range}

\item{logspace}{Flag indicating whether to space test values logarithmically.}
}
\value{
logical
}
\description{
This is a quick and dirty check; it will definitely fail if the input range
is not in the domain of both functions.
}
\keyword{internal}
