% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{concatGrids}
\alias{concatGrids}
\title{Concatenate a list of griddata objects}
\usage{
concatGrids(gridlist)
}
\arguments{
\item{gridlist}{List of grids to concatenate}
}
\value{
A griddata (q.v. \code{link{read.temperatures}}) object with the concatenated
grid values.
}
\description{
Griddata objects are concatenated by appending their temperature observations
into a single grand matrix with all the time sices from the concatenated
objects.
}
\details{
The time indices are likewise concatenated so that they will give accurate
time values.  The tags are concatenated and updated so that they point to the
new location of their original data.  Lat and lon values and the tgop
operator are unchanged, and they must be the same for all of the concatenated
grids, or an error is raised.

Note: Time values are given as years since the start year for the data set.
If you concatenate data sets with different start years, the time values will
not be compatible.  For \emph{most} of what we do in this package, the
difference won't matter, but it's something to be aware of.
}
