% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_NAs.R
\name{coord_array}
\alias{coord_array}
\title{Convert lat/lon coordinates to a coordinate array}
\usage{
coord_array(lat, lon)
}
\arguments{
\item{lat}{Latitude coordinate array for the full rectangular grid}

\item{lon}{Longitude coordinate array for the full rectangular grid}
}
\value{
Matrix \link{ngrid x 2}, latitude in the first column, longitude in the
second.
}
\description{
NetCDF assumes the grid is rectangular, so it stores latitude by row and
longitude by column.  We are potentially handling irregular coordinate grids
here, so we need to be able to covert to lat/lon for each grid cell.
}
