% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_NAs.R
\name{drop_NAs}
\alias{drop_NAs}
\title{Remove NA grid cells from processing in the emulator training process}
\usage{
drop_NAs(griddata_input)
}
\arguments{
\item{griddata_input}{output from \code{concatGrids.general}}
}
\value{
A list similar in structure and content to \code{concatGrids.general} output but without the
NA gridcells if applicable and a mapping data frame of the original and new grid cell indices.
}
\description{
This function removes grid cells with missing data, NA values for all timesteps,
from being processed by \code{trainTP}. This enables \code{fldgen} to work with output that has NAs in it.
}
\keyword{internal}
