% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writedata.R
\name{emulator_reducer}
\alias{emulator_reducer}
\title{Subset a trained emulator.}
\usage{
emulator_reducer(emulator)
}
\arguments{
\item{emulator}{A trained fldgen emulator, with all entries needed
for generating new residuals and for rigourously validating the
quality of the trained emulator}
}
\value{
reducedEmulator A trained fldgen emulator with only the list
entries needed by generate.TP.resids and generate.TP.fullgrids for
generating new fields:
\describe{
\item{griddataT}{Only the coordinate ids and set information.}
\item{griddataP}{Only the coordinate ids and set information, and the
function to convert from logP to P.}\
\item{tgav}{The Tgav data from training.}
\item{meanfldT}{the slope (w) and intercept (b) terms from the mean field
fit.}
\item{meanfldP}{the slope (w) and intercept (b) terms from the mean field
fit.}
\item{tfuns}{The empirical quantile functions for temperature, mapping
N(0,1) to the native distribution in each grid cell.}
\item{pfuns}{The empirical quantile functions for logP, mapping
N(0,1) to the native distribution in each grid cell.}
\item{reof}{The EOFs.}
\item{fx}{Time coefficients for each EOF from training data.}
\item{infiles}{The names of the files used for training the emulator.}
}
}
\description{
A trained fldgen emulator features a large amount of data for both
using the emulator and rigorously validating an emulator.
}
\details{
If one is just interested in the use of an emulator for generating
felds, this function can be called to reduce a trained emulator to
the bare essential list entries, which can then be saved and called
the same as an unreduced emulator by generate.TP.resids and
generate.TP.fullgrids

Note that with this reduced emulator, there is NO way to reconstruct
the training data. A fully trained emulator contains a copy of the
training data, in addition to the training regressor values (tgav),
and the estimated linear model parameters and residuals
(meanfieldT$b, w, r), which together can also reconstruct the data.

Even though the coordinate information stored in an emulator$griddataT
is not needed directly to generate a new field of residuals or full data,
it is often needed in downstream use of the fields. Therfore an entry
reducedEmulator$griddataT$coord containg a matrix is saved in the
reducedEmulator. Each is a matrix of coordinates for each grid cell, with
cells in rows and latitude, longitude in the two columns. Keeping these
coordinate matrices for T and P is negligible.

Finally, the reduced emulator produced by this function is specifically
meant for temperature and precipitation only, and is not robust to
extension to other variables.

Finally finally, if a user is interested in a different subset of
list entries in a trained emulator, they are encouraged to subset and
save themself, as appropriate for their project.
}
\author{
ACS July 2020
}
