% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfield.R
\name{eof_analyze}
\alias{eof_analyze}
\title{Perform EOF analysis on residuals}
\usage{
eof_analyze(resids, tgop)
}
\arguments{
\item{resids}{A matrix [ntime x ngrid] of residuals from the mean field}

\item{tgop}{A column vector [ngrid x 1] containing the global mean operator.
This is produced as the \code{tgop} field of the structure returned by
\code{\link{read.temperatures}}.}
}
\description{
Decompose residuals from the mean field into orthogonal components.  One of
these components (conventionally the first component) is always the
normalized global mean temperature operator.  This convention isolates the
time variability of the global mean in a single components and ensures that
the remaining components describe variability that is purely spatial.
}
\section{Value}{


The return value of this function is in the same format as the structure
returned by \code{\link[stats]{princomp}}.  The most important fields are:

\describe{
  \item{rotation}{A matrix [ngrid x N] containing orthogonal basis vectors
  for the residuals.  Each column is a basis vector.  The number of basis
  vectors, N, will be less than or equal to the number of time slices in the
  input.}
  \item{x}{A matrix [ntime x N] containing the coordinates of the residuals
  in the coordinate system defined by the basis vectors.  Each row is a time
  slice.  Thus, \eqn{r(t) = \sum_{i=1}^{N} x[t,i] * rotation[:,i]}.}
}
}

