% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamunge.R
\name{extract_box}
\alias{extract_box}
\title{Create an operator to extract grid cells from a field by latitude and
longitude.}
\usage{
extract_box(lat, lon, griddata)
}
\arguments{
\item{lat}{A vector of length 2 giving the minimum and maximum latitude for
the bounding box.}

\item{lon}{A vector of length 2 giving the minimum and maximum longitude for
the bounding box.}

\item{griddata}{A \code{griddata} structure, which contains the latitude and
longitude coordinates.}
}
\value{
Logical vector indicating which grid cells are in the bounding box.
}
\description{
This function creates an operator for extracting a selection of grid cells
identified by a latitude and longitude bounding box.  The result is a logical
vector with \code{TRUE} values in the indices corresponding to grid cells
inside the bounding box.
}
\details{
This operator can be used as an index to extract the desired grid cells, for
example:

\code{mask <- extract_box; fld_subset <- fld[ ,mask]}.

It can also
be used to calculate averages over a specified area:

\code{meantemp <- fld \%*\% as.numeric(mask)}.

Or, you could use it to mask out grid cells that are not in the region of
interest:

\code{fld[ ,!mask] <- NA}
}
