% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{file.pairer}
\alias{file.pairer}
\title{Pair temperature and precipitation (or any other variable) netCDF files
based on file name, following CMIP5 convention.}
\usage{
file.pairer(dat, tvarname = "tas", pvarname = "pr")
}
\arguments{
\item{dat}{A single directory name, or a list of netCDF files.  If a
directory name is given, all netCDF files in the directory will be used.The
pairing of temperature and precipitation netCDF files in the directory
relies on the CMIP5 file naming conventions. Other naming conventions are
not currently supported.}

\item{tvarname}{The name of the temperature variable in CMIP5 ('tas' default).}

\item{varname}{The name of the precipitation variable in CMIP5 ('pr' default).}
}
\value{
A tibble of paired file names in two columns: tfilename, pfilename.
Each row of the tibble is a single scenario.
}
\description{
This function reads in a directory or list of netCDF files, with names
following CMIP5 scenario design conventions, and pairs the temperature
and precipitation files based on scenario. Technically any two variables
could have files paired using this function, just by changing the tvarname
and pvarname inputs.
}
\keyword{internal}
