% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{flatten_tags}
\alias{flatten_tags}
\title{Flatten a list of tags, restoring their start/end rows}
\usage{
flatten_tags(taglist)
}
\arguments{
\item{taglist}{List of tags, possibly nested}
}
\description{
Flatten a list of tags, restoring their start/end rows
}
\keyword{internal}
