% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fldgen.R
\docType{package}
\name{fldgen-package}
\alias{fldgen}
\alias{fldgen-package}
\title{Generate temperature fields with spatial and temporal correlation}
\description{
Provides functions to learn the spatial and temporal correlation of the
variability in an ESM and generate random temperature fields with equivalent
properties.
}
\author{
\strong{Maintainer}: Robert Link \email{robert.link@pnnl.gov}

}
