% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{fldgen_object}
\alias{fldgen_object}
\title{Create a \code{fldgen} object from constituent parts}
\usage{
fldgen_object(griddata, tgav, pscl, reof, ftran)
}
\arguments{
\item{griddata}{An object returned from \code{\link{read.temperatures}} or
\code{\link{concatGrids}}.}

\item{tgav}{Global mean temperature for the grids in griddata.}

\item{pscl}{Object returned from \code{\link{pscl_analyze}}.}

\item{reof}{Object returned from \code{\link{eof_analyze}}.}

\item{ftran}{The Fourier transform of the EOF projection coefficients.}
}
\description{
Normally this code will be called internally, but it is available for
external use, so that data created using the old interface can be converted.
}
\keyword{internal}
