% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTP.R
\name{fldgen_object_TP}
\alias{fldgen_object_TP}
\title{Create a \code{fldgen} object from constituent parts}
\usage{
fldgen_object_TP(griddataT, griddataP, tgav, meanfldT, meanfldP, tfuns, pfuns,
  reof, fxmag, fxphase, infiles)
}
\arguments{
\item{griddataT}{An object returned from \code{\link{read.temperatures}} or
\code{\link{concatGrids}}.}

\item{griddataP}{An object returned from \code{\link{read.temperatures}} or
\code{\link{concatGrids}}.}

\item{tgav}{Global mean temperature for the grids in griddata.}

\item{meanfldT}{Object returned from whatever mean field analysis is done,
currently \code{\link{pscl_analyze}}.}

\item{meanfldP}{Object returned from whatever mean field analysis is done,
currently \code{\link{pscl_analyze}}.}

\item{tfuns}{The empirical cdf and empirical quantile functions for each
grid cell, characterizing the distribution of residuals resulting from
meanfldT.}

\item{pfuns}{The empirical cdf and empirical quantile functions for each
grid cell, characterizing the distribution of residuals resulting from
meanfldP.}

\item{reof}{Object returned from \code{\link{eof_analyze}}.}

\item{fxmag}{The magnitude of the Fourier transform of the EOF projection
coefficients.  This should be a matrix [Ntime x NEOF].  If using
\code{\link{psdest}}, note this is the square root of the power spectral
density returned by that function.}

\item{fxphase}{List of matrices [Ntime x NEOF] of phases of the Fourier
transform.  There should be one element in the list for each input ESM run.}

\item{infiles}{Names of input files used to construct the data.}
}
\description{
Normally this code will be called internally, but it is available for
external use, so that data created using the old interface can be converted.
}
\details{
If there is only a single ESM run in the input data set, then \code{fxphase}
can be passed as a matrix of phases (instead of a list of matrices).  It will
be converted into a list automatically.
}
\keyword{internal}
