% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeresiduals.R
\name{normalize.resids}
\alias{normalize.resids}
\title{Normalize residuals}
\usage{
normalize.resids(inputresids, empiricalcdf, len = NULL)
}
\arguments{
\item{inputresids}{Matrix of the original, input residuals for each grid
cell (columns)}

\item{empiricalcdf}{List of the empirical cdf functions for each grid cell.}

\item{len}{Maximum length of the time series to read.  If the data read is
longer, it will be trimmed.  (Default: read entire time series, regardless of
length.)}
}
\description{
A vector of residuals in a single grid cell must be normalized for the
fldgen algorithm to work properly. This function takes a matrix of
residuals (each grid cell is a column, T or P), calculates the quantiles
of each residual, and maps to the corresponding value in a normal
distribution.
}
\details{
The output will be a list with four fields:
\describe{
  \item{\strong{inputresids}}{Matrix of the original, input residuals for
  each grid cell.}
  \item{\strong{empiricalcdf}}{List of the empirical cdf functions for each
  grid cell.}
  \item{\strong{quants}}{Matrix of the quantiles of each input residual.}
  \item{\strong{rn}}{Matrix of the new, normally distributed residuals.}
}



Conventionally, we refer to the output list as \code{quantilemapping}.
Notably, any other function with a \code{quantilemapping} argument is
expecting one of these structures.
}
