% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{phase_eqn_coef}
\alias{phase_eqn_coef}
\title{Generate the coefficients in the equations that determine the phase constraints.}
\usage{
phase_eqn_coef(Fx, i = 2)
}
\arguments{
\item{Fx}{The Fourier transform of the EOF projection coefficients.}

\item{i}{The index of the basis function to use for the reference.  Don't use
EOF-0 (i=1).}
}
\description{
The result will be a structure containing:
\describe{
  \item{A}{A matrix [Nf x Nb], where Nf is the number of
nonnegative frequencies in the Fourier transform and Nb is the number of
basis functions (i.e., EOFs).  A[k,j] = (a_k^i a_k^j), where i
is the index of the basis function used for the reference (conventionally 2,
which is the first spatial basis function.  Don't use the time-like basis
function, i=1, since it will generally have small coefficients).}
  \item{i}{The index of the reference basis function.}
}
}
\details{
These equations are given as equation XX in the paper.
}
\keyword{internal}
