% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_field}
\alias{plot_field}
\title{Plot a single field in matrix form}
\usage{
plot_field(fld, griddata, nb = 6, minval = -3.5, maxval = 3.5)
}
\arguments{
\item{fld}{Vector of ngrid=nlat*nlon values: a single time slice of the
field.}

\item{griddata}{The griddata structure returned from
\code{\link{read.temperatures}}.}

\item{nb}{Number of breaks in the color scale.  If nb < 2, use a smooth
gradient.}

\item{minval}{Lower limit of the color scale.  The default value was chosen
to work well for fields of residuals from the mean temperature response.}

\item{maxval}{Upper limit of the color scale.  The default value was chosen
to work well for fields of residuals from the mean temperature response.}
}
\description{
Transform the field into a data frame using \code{\link{fld2df}} and plot
using the \code{gcammaptools} package.  If \code{gcammaptools} isn't
available, return \code{NULL}.  (TODO: make an ersatz plot if gcammaptools
isn't there)
}
