% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotglobalfieldsP}
\alias{plotglobalfieldsP}
\title{Plot a single year's precipitation for every generated gridded time series
in a list.}
\usage{
plotglobalfieldsP(fieldlist, yearindex, emulatorgriddata, minval = -1.5,
  maxval = 1.5, legendstr, palettestr = "BrBG", palettedir = 1)
}
\arguments{
\item{fieldlist}{List of residual}

\item{yearindex}{Index of the year to be plotted. For example, in data
ranging from 2006 to 2100, the yearindex for 2006 is 1.}

\item{emulatorgriddata}{The griddataP structure returned in a trained
emulator.}

\item{minval}{Lower limit of the color scale.  The default value was chosen
to work well for fields of residuals from the mean precipitation response.}

\item{maxval}{Upper limit of the color scale.  The default value was chosen
to work well for fields of residuals from the mean precipitation response.}

\item{legendstr}{String to use for the title of the graph legend}

\item{palettestr}{String with the name of the ggplot2 color palette to use.
Defaults to 'RdYlBu'. Options are Diverging: 'BrBG', 'PiYG', 'PRGn', 'PuOr',
'RdBu', 'RdGy', 'RdYlBu', 'RdYlGn', 'Spectral'; Qualitative: 'Accent',
'Dark2', 'Paired', 'Pastel1', 'Pastel2', 'Set1', 'Set2', 'Set3';
Sequential: 'Blues', 'BuGn', 'BuPu', 'GnBu', 'Greens', 'Greys', 'Oranges',
'OrRd', 'PuBu', 'PuBuGn', 'PuRd', 'Purples', 'RdPu', 'Reds', 'YlGn',
'YlGnBu', 'YlOrBr', 'YlOrRd'.}

\item{palettedir}{Numerical value, only accepts 1 or -1, determining the
order of the color palette used. Defaults to -1.}
}
\description{
Takes in a list of precipitation fields (residual or full data), a year
index, and the underlying precipitation gridded data from a trained
emulator, as well as a few arguments relevant to plotting
}
\author{
ACS May 2020
}
