% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfield.R
\name{pscl_analyze}
\alias{pscl_analyze}
\title{Run pattern scaling analysis on the input temperature fields}
\usage{
pscl_analyze(tin, tgavin)
}
\arguments{
\item{tin}{Matrix of temperature fields (ntime x ngrid)}

\item{tgavin}{Matrix of global mean temperature over time (ntime x 1)}
}
\value{
List of r, w, and b matrices (see details)
}
\description{
Run pattern scaling on the input temperature fields.  This amounts to doing a
linear regression on each grid cell, with the global temperature as the sole
regressor.
}
\details{
The return value is a list with the residuals and pattern coefficients as r
(resid), w (slope), and b (intercept).

dim(r) == dim(tin) == ntime x nlat*nlon
dim(w) == dim(b) == nlat*nlon
}
\keyword{internal}
