% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfield.R
\name{pscl_apply}
\alias{pscl_apply}
\title{Generate a mean field from a pattern and an input temperature series}
\usage{
pscl_apply(pscl, tgav)
}
\arguments{
\item{pscl}{A scaling pattern generated by \code{\link{pscl_analyze}}.}

\item{tgav}{A [N x 1] matrix of global mean temperatures.}
}
\description{
Apply the w and b values to the input global mean temperatures.  Thus, for
each grid cell \eqn{i}, \deqn{T_i = w_i T_g + b_i}.  The residuals are
\emph{not} added back to the result.
}
\details{
The input temperatures should be in the form of a column vector; i.e., a
matrix with dimension N x 1.  The output will be a matrix of dimension N x
ngrid, where each row is the flattened vector of grid cells for a single time
slice.

For purposes of applying the pattern, any value of N will work; however, many
of the other functions in this package expect a matrix with dimension ntime x
ngrid, where ntime is the number of time slices in the ESM dataset originally
used in the analysis.  Therefore, for any output that is intended to be used
in these other functions, the input temperatures should be of length ntime.
}
\keyword{internal}
