% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{read.ncdf}
\alias{read.ncdf}
\title{Read ESM data}
\usage{
read.ncdf(filename, len = NULL, varname = "tas")
}
\arguments{
\item{filename}{Name of the input netCDF file}

\item{len}{Maximum length of the time series to read.  If the data read is
longer, it will be trimmed.  (Default: read entire time series, regardless of
length.)}

\item{varname}{Name of the variable to read from the netcdf file.
Irrespective of the name in the netcdf file, it will be called 'tas' in the
return data.}
}
\value{
A \code{griddata} list (see details).
}
\description{
Read temperature data from the ESM that we are going to emulate.
}
\details{
The temperature data should be a netCDF file of annual grid values.  The data
is assumed to have its dimensions of time, lat, lon, where lon is the most
rapidly varying index, and time is the least.

The output will be a list with two fields:
\describe{
  \item{\strong{tas}}{Matrix of temperature data.}
  \item{\strong{tgop}}{Vector operator for global mean temperature.}
}

The data at each time is represented as a flattened vector of grid cells.
The flattening is performed by transposing to lon, lat ordering, so that lon
will once again be the most rapidly varying index (because R uses
Fortran-style indexing).  Then the spatial dimensions are discarded,
resulting in a 1D vector.  The time dimension is kept, resulting in a matrix
with years in rows and grid cells in columns.  The dimensions of the matrix
will be Nyear x Ngrid, where Nyear is the number of years in the data set,
and Ngrid is the number of grid cells.

The \code{tgop} vector is a vector whose dot product with a flattened grid is
the area-weighted global mean for the grid.  It is stored as a column vector,
so \code{tas \%*\% tgop} is the time series of global mean temperatures.

The lat and lon dimension variables from the input file  are also
stored in the structure.  These are primarily useful for writing out
generated grids as netCDF files.  The time dimension variable is converted to
integer years, starting at 0.

Conventionally, we refer to the output list as \code{griddata}.  Notably, any
other function with a \code{griddata} argument is expecting one of these
structures.
}
