% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{readdata}
\alias{readdata}
\title{Read ESM data}
\usage{
readdata(filename)
}
\arguments{
\item{filename}{Name of the input netCDF file}
}
\value{
A \code{griddata} list (see details).
}
\description{
Read temperature data from the ESM that we are going to emulate.
}
\details{
The temperature data should be a netCDF file of annual grid values.  The data
is assumed to have its dimensions of time, lat, lon, where lon is the most
rapidly varying index, and time is the least.

The output will be a list with two fields:
\describe{
  \item{\strong{tas}}{Matrix of temperature data.}
  \item{\strong{tgop}}{Vector operator for global mean temperature.}
}

The data at each time is represented as a flattened vector of grid cells.
The flattening is performed by transposing to lon, lat ordering, so that lon
will once again be the most rapidly varying index (because R uses
Fortran-style indexing).  Then the lat/lon dimensions are discarded,
resulting in a 1D vector.  The time dimension is kept, resulting in a matrix
with years in rows and grid cells in columns.  The dimensions of the matrix
will be Nyear x Ngrid, where Nyear is the number of years in the data set,
and Ngrid is the number of grid cells.

The \code{tgop} vector is a vector whose dot product with a flattened grid is
the area-weighted global mean for the grid.  It is stored as a column vector,
so \code{tas \%*\% tgop} is the time series of global mean temperatures.

Conventionally, we refer to the output list as \code{griddata}.  Notably, any
other function with a \code{griddata} argument is expecting one of these
structures.
}
