% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{readtgav}
\alias{readtgav}
\title{Read and format global mean temperature}
\usage{
readtgav(tgavfilename)
}
\arguments{
\item{tgavfilename}{netCDF file with global mean temperatures variable}
}
\value{
A ntime x 1 matrix of global mean temperature
}
\description{
Read global mean temperature from an input netCDF file and format for use
with the rest of the processing functions.
}
\details{
Global mean temperature is a 1-D array, so storing in a netCDF file isn't
strictly necessary; however, some tools produce it in that format by
default.  A vector of tgav values from any source can be used with the other
functions in this package; all you have to do is convert it to an ntime x 1
matrix using \code{as.matrix}.
}
