% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfield.R
\name{reconst_fields}
\alias{reconst_fields}
\title{Reconstitute temperature fields}
\usage{
reconst_fields(basis, bcoord, meanfield = NULL)
}
\arguments{
\item{basis}{A matrix [ngrid x N] of basis vectors, one vector per column.}

\item{bcoord}{A matrix [ntime x N] of basis coordinates, one time step in
each row, with coordinates across the columns.}

\item{meanfield}{An optional matrix [ntime x ngrid] of mean field values, one
time slice per row. If supplied, it will be added to the residuals calculated
from \code{basis} and \code{bcoord}.}
}
\value{
A matrix [ntime x ngrid] of residual values (or temperature values,
if \code{meanfield} was supplied).
}
\description{
Reconstitute temperature residual fields from the basis vectors and a
matrix of coefficients.  If meanfield is supplied, add the residual field
to the mean field to get a reconstituted temperature field.
}
