% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R, R/writedata.R
\name{loadmodel}
\alias{loadmodel}
\alias{saving_and_restoring}
\alias{savemodel}
\title{Load and save emulator training data}
\usage{
loadmodel(file)

savemodel(modeldata, file, clobber = FALSE)
}
\arguments{
\item{file}{Name of the file to write the data to.}

\item{modeldata}{A \code{fldgen} object returned by either
\code{\link{train}} or \code{\link{fldgen_object}}.}

\item{clobber}{Flag indicating whether it's ok to overwrite an existing file}
}
\description{
\code{savemodel} saves the results of training an emulator in a portable
format.  \code{loadmodel} loads a model from a file created this way and
returns it as a \code{fldgen} object.
}
