% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{splitGrids.general}
\alias{splitGrids.general}
\title{Split a concatenated griddata object into its constituent grids}
\usage{
splitGrids.general(griddata)
}
\arguments{
\item{griddata}{A griddata structure to split}
}
\value{
A list of griddata structures, each containing a single data set
(i.e., a time series read in from a single file)
}
\description{
This function undoes the effect of \code{\link{concatGrids}}.  It does not,
however, remember the path you took to get the concatenated grid, so
\code{splitGrids(c(g1, c(g2,g3)))} will give you \code{list(g1,g2,g3)}
instead of \code{list(g1, c(g2,g3))}, as you might have expected.
}
\keyword{internal}
