% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_fldgen}
\alias{train_fldgen}
\title{Train the emulator on a set of input ESM data.}
\usage{
train_fldgen(dat, latvar = "lat", lonvar = "lon")
}
\arguments{
\item{dat}{A single directory name, or a list of netCDF files.  If a
directory name is given, all netCDF files in the directory will be used.}

\item{latvar}{Name of the latitude coordinate variable in the netCDF files.
Normally this is \code{'lat'}, but occasionally it will be something
different, such as \code{'lat_2'}.}

\item{lonvar}{Name of the longitude coordinate variable in the netCDF files.}
}
\value{
A \code{fldgen} object.
}
\description{
This function runs all of the steps in the analysis that depend only on the
input data.  The results are saved in a structure of class \code{fldgen}.
This structure contains everything the emulator has learned about the model.
Therefore, it can be saved and reloaded in future sessions to bypass the
analysis step.
}
\details{
The steps run by this procedure are \code{\link{read.temperatures}},
\code{\link{pscl_analyze}}, and \code{\link{eof_analyze}}.  Additionally, the
global mean temperature and the FFT of the EOF decomposition are calculated
and stored.  For backward compatibility, all of those functions remain part
of the public interface.

At present the field generation functions still take as arguments the
individual components of the structure; however, in the next version they
will expect a \code{fldgen} object, which can be constructed manually using
\code{\link{fldgen_object}}.
}
