% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writedata.R
\name{write.temperature}
\alias{write.temperature}
\title{Write a temperature field as a netcdf file.}
\usage{
write.temperature(
  fld,
  file,
  griddata,
  varname = "tas",
  varunit = "K",
  vardesc = "2m air temperature",
  tunit = "years since 2006",
  clobber = FALSE
)
}
\arguments{
\item{fld}{A matrix [ntime x ngrid] containing the field to write out.}

\item{file}{Name of the output file.}

\item{griddata}{A griddata structure returned from the
\code{\link{read.temperatures}} function.}

\item{varname}{Name to use for the variable in the output file.}

\item{varunit}{Units of the variable in the field.}

\item{vardesc}{Description string to write into the \code{longname} attribute
in the output file.}

\item{tunit}{Units for the time variable.}

\item{clobber}{If \code{TRUE}, overwrite the file if it exists; otherwise
trying to write  a file that already exists is an error.}
}
\description{
Format a field as a netcdf file and write it to the specified file.  The lat,
lon, and time variables from the griddata input are used to recover the
original size of the grid, reversing the flattening procedure.
}
