% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_NAs.R
\name{add_NAs}
\alias{add_NAs}
\title{Add the NA grid cells that may have been removed during the the emulator training process}
\usage{
add_NAs(data, griddata)
}
\arguments{
\item{data}{Matrix of data to which to restore missing grid cells}

\item{griddata}{Griddata structure from the training procedure}
}
\value{
Matrix of data with grid cells restored.
}
\description{
This function adds NA gridcells that were removed during the training process in order to convert
the results to same dimensions as the input grid.
}
\keyword{internal}
