% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fldgen.R
\docType{package}
\name{fldgen-package}
\alias{fldgen}
\alias{fldgen-package}
\title{Generate temperature fields with spatial and temporal correlation}
\description{
Provides functions to learn the spatial and temporal correlation of the
variability in an ESM and generate random temperature fields with equivalent
properties.
}
\section{Example}{


The data used in this example can be found in \code{system.file('extdata', package='fldgen')}

\verb{
library(fldgen)

emulator <- train('extdata/tann1.nc')

tempgrids <- list()                     # Empty list to hold the temperature
                                       # realizations we are about to create.
length(tempgrids) <- 4

##  Run with the phases of the actual time series
meanfield <- pscl_apply(emulator$pscl, emulator$tgav)
tempgrids[[1]] <- reconst_fields(emulator$reof$rotation, mkcorrts(emulator) , meanfield)
## Run the rest with random phases
for(i in 2:4)
    tempgrids[[i]] <- reconst_fields(emulator$reof$rotation, mkcorrts(emulator), meanfield)
}
}

\author{
\strong{Maintainer}: Robert Link \email{robert.link@pnnl.gov}

}
