% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTPfullgrids.R
\name{generate.TP.fullgrids}
\alias{generate.TP.fullgrids}
\title{Using a trained emulator, generate new full fields.}
\usage{
generate.TP.fullgrids(emulator, residgrids, tgav,
  reconstruction_function = pscl_apply, addNAs = FALSE)
}
\arguments{
\item{emulator}{A trained \code{fldgen} temperature precipitation joint
emulator.}

\item{residgrids}{Descript}

\item{tgav}{Descript}

\item{reconstruction_function}{Descript}

\item{addNAs}{descript}
}
\value{
Descript - can crib some of the language from generateTPresids
}
\description{
This function takes in a trained emulator - a structure of class
\code{fldgen}.
This structure contains everything the emulator has learned about the model,
and is used to generate new fields of residuals.
}
\details{
First, ...
}
