% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTPfullgrids.R
\name{generate.TP.fullgrids}
\alias{generate.TP.fullgrids}
\title{Using a trained emulator, generate new full fields.}
\usage{
generate.TP.fullgrids(emulator, residgrids, tgavdf, tvarunconvert_fcn = NULL,
  pvarunconvert_fcn = exp, reconstruction_function = pscl_apply)
}
\arguments{
\item{emulator}{A trained \code{fldgen} temperature precipitation joint
emulator.}

\item{residgrids}{A list of new residual fields, each entry in the list is a
new realization, a matrix that is [Nyears x 2 * Ngrid]; the first 1:Ngrid
cols are the temperature residuals and columns (Ngrid + 1):(2*Ngrid) are the
precipitation residuals.}

\item{tgavdf}{A data frame with two columns. Column \code{tgav} = the vector
of global annual mean temperatures for constructing a mean field with
\code{reconstruction_function}.  To be added to each list entry of residual
fields in \code{residgrids}. And a column \code{time}, used to subset the
trained emulator so that pattern scaling information from the appropriate
years is applied to tgav time series. Note: grids cannot be generated for
tgav years missing from the trained emulator.}

\item{tvarunconvert_fcn}{The function to undo any transformation done to the
input training data in \code{trainTP} to correct the support. This should be
the inverse function of the tvarconvert_fcn argument to \code{trainTP}. This
is stored in a \code{trainTP} \code{emulator} under
\code{emulator$griddattaT$tvarconvert_fcn}.
Defaults to NULL as temperature typically doesn't need to be transformed to
correct the support. WARNING: currently rely on the user to define the
correct inverse function here, though we do some checks and throw warnings
if it looks like there may be an issue.}

\item{pvarunconvert_fcn}{The function to undo any transformation done to the
input training data in \code{trainTP} to correct the support. This should be
the inverse function of the pvarconvert_fcn argument to \code{trainTP}. This
is stored in a \code{trainTP} \code{emulator} under
\code{emulator$griddattaP$pvarconvert_fcn}.
Defaults to exp() as precipitation is usually log-transformed in ordere to
correct the support. WARNING: currently rely on the user to define the
correct inverse function here, though we do some checks and throw warnings
if it looks like there may be an issue.}

\item{reconstruction_function}{A function for constructing a mean field from
trained pattern scaling result + a vector of global annual mean temperatures.}
}
\value{
A list of:
1) fullgrids = A list of new full fields, each entry in
the list is a new realization, a matrix that is [Nyears x 2 * Ngrid]; the
first 1:Ngrid cols are the temperature field and columns
(Ngrid + 1):(2*Ngrid) are the precipitation field.
2) time = the input time column used to subset the trained emulator so that
pattern scaling information from the appropriate years is applied to tgav
time series. Note: grids cannot be generated for tgav years missing from
the trained emulator.
3) meanfieldT = the reconstructed, pattern scaled temperature mean field.
4) meanfieldP = the reconstructed, pattern scaled precipitation mean field.
}
\description{
This function takes in a trained emulator - a structure of class
\code{fldgen}.
This structure contains everything the emulator has learned about the model,
and is used to generate new fields of residuals. Also taking in a list of
generated residual fields and a global average yield, a global gridded
mean field is constructed accoring to the input reconstruction_function.
The mean field and residual fields are added to return a list of different
realizations of full fields.
}
