% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTPfullgrids.R
\name{generate.TP.fullgrids}
\alias{generate.TP.fullgrids}
\title{Using a trained emulator, generate new full fields.}
\usage{
generate.TP.fullgrids(emulator, residgrids, tgav,
  reconstruction_function = pscl_apply, addNAs = FALSE,
  tvarunconvert_fcn = NULL, pvarunconvert_fcn = exp)
}
\arguments{
\item{emulator}{A trained \code{fldgen} temperature precipitation joint
emulator.}

\item{residgrids}{Descript}

\item{tgav}{Descript}

\item{reconstruction_function}{Descript}

\item{addNAs}{descript}

\item{tvarunconvert_fcn}{The function to undo any transformation done to the
input training data in \code{trainTP} to correct the support. This should be
the inverse function of the tvarconvert_fcn argument to \code{trainTP}.
Defaults to NULL as temperature typically doesn't need to be transformed to
correct the support. WARNING: currently rely on the user to define the
correct inverse function here.}

\item{pvarunconvert_fcn}{The function to undo any transformation done to the
input training data in \code{trainTP} to correct the support. This should be
the inverse function of the pvarconvert_fcn argument to \code{trainTP}.
Defaults to exp() as precipitation is usually log-transformed in ordere to
correct the support. WARNING: currently rely on the user to define the
correct inverse function here.}
}
\value{
Descript - can crib some of the language from generateTPresids
}
\description{
This function takes in a trained emulator - a structure of class
\code{fldgen}.
This structure contains everything the emulator has learned about the model,
and is used to generate new fields of residuals.
}
\details{
First, ...
}
