% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTP.R
\name{trainTP}
\alias{trainTP}
\title{Train the emulator on a set of input ESM temperature and precipitation data.}
\usage{
trainTP(dat, tvarname = "tas", tlatvar = "lat", tlonvar = "lon",
  tvarconvert_fcn = NULL, pvarname = "pr", platvar = "lat",
  plonvar = "lon", pvarconvert_fcn = logPfloor,
  meanfield = pscl_analyze, globalAvg_file = NULL,
  record_absolute = FALSE)
}
\arguments{
\item{dat}{A single directory name, or a list of netCDF files.  If a
directory name is given, all netCDF files in the directory will be used.The
pairing of temperature and precipitation netCDF files in the directory
relies on the CMIP5 file naming conventions. Other naming conventions are
not currently supported.}

\item{tvarname}{Name of the temperature variable in the temperature netCDF}

\item{tlatvar}{Name of the latitude coordinate variable in the temperature
netCDF files.
Normally this is \code{'lat'}, but occasionally it will be something
different, such as \code{'lat_2'}. Can vary between paired T and P.}

\item{tlonvar}{Name of the longitude coordinate variable in the temperature
netCDF files.  Can vary between paired T and P.}

\item{tvarconvert_fcn}{The function used to transform the T variable prior
to training so that it has support on -infinity to infinity. Defaults to
NULL, as Temperature is effectively already supported on this range.}

\item{pvarname}{Name of the precipitation variable in the precipitation
netCDF.}

\item{platvar}{Name of the latitude coordinate variable in the precipitation
netCDF files.
Normally this is \code{'lat'}, but occasionally it will be something
different, such as \code{'lat_2'}.  Can vary between paired T and P.}

\item{plonvar}{Name of the longitude coordinate variable in precipitation
netCDF files.  Can vary between paired T and P.}

\item{pvarconvert_fcn}{The function used to transform the P variable prior
to analyis so that it has support on -infinity to infinity. Defaults to
log() as precipitation values cannot be less than 0.}

\item{meanfield}{Function to compute the mean temperature response field.
The default is a linear pattern scaling algorithm.}

\item{globalAvg_file}{Optional string to use in constructing the name of the
file containing global mean temperatures, Tgav.  If omitted, calculate Tgav
as the mean of the grid cells in each frame. (See "Supplying global mean
temperatures" for more information.)}

\item{record_absolute}{If \code{TRUE}, record absolute paths for the input
files; otherwise, record relative paths.}
}
\value{
A \code{fldgen} object.
}
\description{
This function runs all of the steps in the analysis that depend only on the
input data.  The results are saved in a structure of class \code{fldgen}.
This structure contains everything the emulator has learned about the model.
Therefore, it can be saved and reloaded in future sessions to bypass the
analysis step.
}
\details{
The steps run by this procedure are \code{\link{read.general}},
\code{\link{pscl_analyze}}, and \code{\link{eof_analyze}}.  Additionally,
the global mean temperature, global mean precipitation, and the FFT of the
EOF decomposition are calculated and stored.  For backward compatibility,
all of those functions remain part of the public interface.

At present the field generation functions still take as arguments the
individual components of the structure; however, in the next version they
will expect a \code{fldgen} object, which can be constructed manually using
\code{\link{fldgen_object}}.

The names of the input files are recorded in the \code{infiles} field of the
returned object.  This is for the your information only; none of the code
here does anything with the recorded filenames.  You have the choice of
recording either absolute filenames (useful for determining exactly which
dataset was used, if you have a lot of similarly named files) or relative
filenames (useful if you might distribute a saved model to other users who
might not have the same directory structure as you).
}
\section{Supplying global mean temperatures}{


In some cases it may be desirable to supply global mean temperatures, rather
than calculating them from the grid.  For example, the grid may cover only
part of the grid, so that the average of the values would not be a true
global average.  In such cases the true global mean temperatures can be
supplied in a separate file.

If this input file is used, it should be a plain text file with one
temperature value per line, each line corresponding to one time slice in the
corresponding netCDF file containing the grid data.  The name of the file
will be constructed by adding a suffix (specified in the
\code{globalAvg_file} argument) to the name of the netCDF file.  For example,
if the netCDF file is called \code{larry.nc}, and the suffix is
\code{curly.txt}, the name of the global mean temperature file will be
\code{larry.nccurly.txt}.  If multiple netCDF files are passed in, each one
will have its own file of global mean temperatures, constructed in the same
way.
}

